fix: production readiness improvements and memory bounds
- Add bounded LRU eviction to RateLimiter HashMaps
  - max_tracked_nodes and max_tracked_ips config options
  - Automatic eviction of oldest entries when capacity reached
  - Monitoring methods for tracked node/IP counts

- Wire ContentStore size tracking to use actual stats
  - get_total_size() now returns get_stats().total_bytes

- Add ContentStoreExtensions tests
  - test_content_store_get_total_size
  - test_content_store_with_strategy

- Fix clippy warnings
  - Use += operator pattern in witness_protocol.rs
  - Use contains() in witness_selection.rs

- Apply cargo fmt to all modified files
- Add patches/ to gitignore

All 794 tests pass with zero clippy warnings.

+++ tests/presence_tests.rs
@@ -19,7 +19,7 @@ fn valid_four_words(seed: u16) -> [String; 4] {
         (seed & 0xFF) as u8,
         (seed % 200) as u8,
     );
-    let port = 15000 + seed as u16;
+    let port = 15000 + seed;
     let encoding = encoder.encode_ipv4(ip, port).expect("encoder works");
     let w = encoding.words();
     [w[0].clone(), w[1].clone(), w[2].clone(), w[3].clone()]
